Time: 3 ms (93.06%), Space: 73 MB (24.49%) - LeetHub

+++ 1803-average-waiting-time/1803-average-waiting-time.java
@@ -1,7 +1,7 @@
 class Solution {
     public double averageWaitingTime(int[][] cust) {
         int ft=0;
-        long avg=0;
+        double avg=0;
         for(int i=0;i<cust.length;i++){
             if(ft<cust[i][0]){
                 ft=cust[i][0];
@@ -9,6 +9,6 @@ class Solution {
             ft+=cust[i][1];
             avg+=ft-cust[i][0];
         }
-        return (double)avg/cust.length;
+        return avg/cust.length;
     }
 }
